% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{descendants_IC}
\alias{descendants_IC}
\title{Get information content based on number of descendants each term has}
\usage{
descendants_IC(ontology)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}
}
\value{
Numeric vector of information contents named by term.
}
\description{
Calculate information content of terms based on frequency with which it is an ancestor of other terms. Useful as a default if there is no population frequency information available as it captures the structure of the ontology.
}
