% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-set-sim.R
\name{get_asym_sim_grid}
\alias{get_asym_sim_grid}
\title{Get asymmetrical similarity matrix}
\usage{
get_asym_sim_grid(A, B, ...)
}
\arguments{
\item{A}{List of term sets.}

\item{B}{List of term sets.}

\item{...}{Other arguments to be passed to \code{\link{get_sim_grid}}.}
}
\value{
Numeric matrix of similarities
}
\description{
Create a numeric matrix of similarities between two lists of term sets, but only averaging over the terms in sets from \code{A} the similarities of the best matches in sets from \code{B}.
}
\seealso{
\code{\link{get_sim_grid}} \code{\link{get_profile_sims}}
}

