% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-set-sim.R
\name{get_profile_sims}
\alias{get_profile_sims}
\title{Get similarities of term sets to profile}
\usage{
get_profile_sims(profile, term_sets, ...)
}
\arguments{
\item{profile}{Character vector of term IDs.}

\item{term_sets}{List of character vectors of ontological term IDs.}

\item{...}{Other arguments to pass to \code{\link{get_sim_grid}}.}
}
\value{
Numeric vector of profile similarities.
}
\description{
Get numeric vector of similarities between each item in a list of term sets and another `ontological profile', i.e. a single term set. Similarity averaging over terms in \code{term_sets}.
}
\seealso{
\code{\link{get_asym_sim_grid}} \code{\link{get_sim_grid}}
}

