% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sim_sample}
\alias{get_sim_sample}
\title{Get random sample of group similarities for groups of given size}
\usage{
get_sim_sample(sym_sim_mat, group_size, sample_size)
}
\arguments{
\item{sym_sim_mat}{Symmetric numeric similarity matrix}

\item{group_size}{Integer giving the number of members of a group}

\item{sample_size}{Number of samples to draw}
}
\value{
Numeric vector of random group similarities
}
\description{
Get random sample of group similarities for groups of given size
}

