% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_similarity_rank_matrix}
\alias{get_similarity_rank_matrix}
\title{Get matrix of similarity rank from similarity matrix}
\usage{
get_similarity_rank_matrix(similarity_matrix)
}
\arguments{
\item{similarity_matrix}{Lower triangular numeric matrix of similarities, where the rownames and colnames are identical to the case IDs}
}
\value{
Rank matrix (not necessarily symmetrical
}
\description{
Given a lower triangular similarity matrix, construct a distance matrix where the rows are the ranks of the column cases, ranked by similarity to the row case
}

