% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lin}
\alias{lin}
\title{Calculate Lin similarity score of two term sets}
\usage{
lin(ontology, information_content, term_set_1, term_set_2)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{information_content}{Numeric vector of information contents of terms (named by term)}

\item{term_set_1}{Character vector of terms}

\item{term_set_2}{Character vector of terms}
}
\description{
Warning! This function is slow - performing large numbers of `between term-set' similarity calculations should be done using \code{\link{get_sim_grid}}.
}
\references{
Lin, D. (1998). An information-theoretic definition of similarity. In Shavlik, J. W., ed., Proceedings of the Fifteenth International Conference on Machine Learning (ICML 1998), Madison, Wisconsin, USA, July 24-27, 1998. (Morgan Kaufmann) pp. 296-304.
}
\seealso{
\code{\link{resnik}}, \code{\link{get_term_sim_mat}}
}

