\name{ontologySimilarity-package}
\alias{ontologySimilarity-package}
\alias{ontologySimilarity}
\docType{package}
\title{
Functions for Calculating Ontological Similarities
}
\description{
Functions for calculating semantic similarities between ontological terms or sets of ontological terms based on term information content and assessing statistical significance of similarity in the context of a collection of phenotypes.
}
\details{
\tabular{ll}{
Package: \tab ontologySimilarity\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-01-12\cr
License: \tab GPL (>= 2) \cr
}
Functions for calculating ontological similarity. Important functions include 
}
\author{
Daniel Greene <dg333@cam.ac.uk>

Maintainer: Daniel Greene <dg333@cam.ac.uk>
}
\references{
	`The Human Phenotype Ontology project: linking molecular biology and disease through phenotype data', Nucl. Acids Res. (1 January 2014) 42 (D1): D966-D974 doi:10.1093/nar/gkt1026
}
\keyword{ ontology, HPO, phenotypic similarity }
