% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-set-sim.R
\name{resnik}
\alias{resnik}
\title{Calculate Resnik similarity score of two term sets}
\usage{
resnik(ontology, information_content, term_set_1, term_set_2)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{information_content}{Numeric vector of information contents of terms (named by term)}

\item{term_set_1}{Character vector of terms.}

\item{term_set_2}{Character vector of terms.}
}
\value{
Numeric value.
}
\description{
Warning! This function is slow - performing large numbers of `between term-set' similarity calculations should be done using \code{\link{get_sim_grid}}.
}
\references{
Resnik, P. (1995). `Using information content to evaluate semantic similarity in a taxonomy'. Proceedings of the 14th IJCAI 1, 448-453.
}
\seealso{
\code{\link{lin}}, \code{\link{get_term_sim_mat}}
}
