% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sim}
\alias{get_sim}
\title{Calculate the group similarity of a set of row/column indices}
\usage{
get_sim(sym_sim_mat, group)
}
\arguments{
\item{sym_sim_mat}{Symmetric numeric similarity matrix}

\item{group}{Integer/character vector giving the indices/names of rows/columns in the \code{sym_sim_mat} for which to compute a similarity}
}
\value{
Numeric value of group similarity
}
\description{
Calculate the group similarity of a set of row/column indices
}

