% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-set-sim.R
\name{get_similarity_rank_matrix}
\alias{get_similarity_rank_matrix}
\title{Get matrix of similarity rank from similarity matrix}
\usage{
get_similarity_rank_matrix(similarity_matrix, symmetric = TRUE)
}
\arguments{
\item{similarity_matrix}{Lower triangular numeric matrix of similarities, where the rownames and colnames are identical to the case IDs.}

\item{symmetric}{Logical value determining whether to `symmetrify' resultant matrix by averaging rank similarity of A -> B and B -> A.}
}
\value{
Matrix of rank similarities.
}
\description{
Given a lower triangular similarity matrix, construct a distance matrix where the rows are the ranks of the column cases with respect to similarity to the row case. If relative similarity is of interest, this rank-transformation may reduce bias in favour of high similarity scores in downstream analysis.
}

