% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-set-sim.R
\name{get_term_set_to_term_sims}
\alias{get_term_set_to_term_sims}
\title{Get `term sets to term' similarity matrix}
\usage{
get_term_set_to_term_sims(term_sets, terms, ...)
}
\arguments{
\item{term_sets}{List of character vectors of ontological term IDs.}

\item{terms}{Character vector of ontological terms.}

\item{...}{Other arguments to be passed to \code{\link{get_sim_grid}}.}
}
\value{
Numeric matrix of term set-to-term similarities
}
\description{
Create a numeric matrix of similarities between term sets and individual terms.
}
\seealso{
\code{\link{get_sim_grid}}
}
